requirements.txt
djangorestframework==3.7.0
django-codemirror2==0.2
django-jsonfield==1.0.1
django-yamlfield==1.0.3
PyYAML==3.12
ruamel.yaml>=0.13.5
simplejson==3.11.1
six==1.11.0
